// src/utils/enums.ts
// eslint-disable-next-line import/prefer-default-export
export enum ResponseCodes {
  NOT_FOUND = 'NOT_FOUND',
  UNAUTHORIZED = 'UNAUTHORIZED',
  UNAUTHENTICATED = 'UNAUTHENTICATED',
  SERVER_ERROR = 'SERVER_ERROR',
  BAD_REQUEST = 'BAD_REQUEST',
  DUPLICATE_DATA = 'DUPLICATE_DATA',
  SUCCESS = 'SUCCESS',
}